@@ -1,10 +1,12 @@
 library squilder.schema_generator;
 
-import 'dart:io';
 import 'dart:async';
-import 'package:squilder/utils.dart';
+import 'dart:ffi';
+import 'dart:io';
+import 'dart:typed_data';
+
 import 'package:squilder/src/generator/generator_config.dart';
-import 'package:dapter/dapter.dart';
+import 'package:squilder/utils.dart';
 
 class _Field {
   final String name;
@@ -13,7 +15,9 @@ class _Field {
   final String key;
   final dynamic defaultValue;
   final String extra;
-  _Field(this.name, this.type, this.isNull, this.key, this.defaultValue, this.extra);
+
+  _Field(this.name, this.type, this.isNull, this.key, this.defaultValue,
+      this.extra);
 
   String toString() {
     return "<_Field name: $name, type: $type, isNull: $isNull, key: $key, defaultValue: $defaultValue, extra: $extra>";
@@ -24,7 +28,9 @@ class _Field {
       return "int";
     } else if (type.startsWith(new RegExp(r"(tiny|small|medium|big)?blob"))) {
       return "List<int>";
-    } else if (type.startsWith("varchar") || type.startsWith(new RegExp(r"(tiny|small|medium|big)?text")) || type.startsWith("char")) {
+    } else if (type.startsWith("varchar") ||
+        type.startsWith(new RegExp(r"(tiny|small|medium|big)?text")) ||
+        type.startsWith("char")) {
       return "String";
     } else if (type.startsWith("date") || type.startsWith("timestamp")) {
       return "DateTime";
@@ -40,14 +46,15 @@ class _Field {
   String get camelizedName => camelize(name);
 }
 
-String _generateCodeForFieldDeclarations(Iterable<_Field> fields, {int shift: 2}) {
+String _generateCodeForFieldDeclarations(Iterable<_Field> fields,
+    {int shift: 2}) {
   return fields.map((field) {
     return """TableField<${field.dartType}> _${field.camelizedName};
 ${" " * shift}TableField<${field.dartType}> get ${field.camelizedName} => _${field.camelizedName};""";
   }).join("\n\n${" " * shift}");
 }
 
-String sqlValueToDart(Object value) {
+String sqlValueToDart(Object? value) {
   if (value is num) {
     return "$value";
   } else if (value == "NULL" || value == null) {
@@ -57,7 +64,8 @@ String sqlValueToDart(Object value) {
   }
 }
 
-String _generateCodeForFieldInitializations(Iterable<_Field> fields, {int shift: 2}) {
+String _generateCodeForFieldInitializations(Iterable<_Field> fields,
+    {int shift: 2}) {
   return fields.map((field) {
     return """_${field.camelizedName} = new TableField<${field.dartType}>(table, "${field.name}", ${sqlValueToDart(field.defaultValue)});""";
   }).join("\n${" " * shift}");
@@ -67,7 +75,8 @@ String _generateCodeForTable(String tableName, Iterable<_Field> fields) {
   var tableClass = "${camelize(tableName, capitalizeFirst: true)}Table";
   var fieldsClass = "${tableClass}Fields";
   var tableNameVar = camelize(tableName);
-  var primaryKeyField = fields.firstWhere((f) => f.name == "id", orElse: () => fields.first);
+  var primaryKeyField =
+      fields.firstWhere((f) => f.name == "id", orElse: () => fields.first);
   var allFields = fields.map((f) => f.camelizedName).join(", ");
   return """class $tableClass extends Table {
   String get name => "$tableName";
@@ -95,18 +104,29 @@ class $fieldsClass extends TableFields {
 }""";
 }
 
-Future<Null> generate({
-    String dbType: "mysql",
-    String host: "localhost",
-    String user: "root",
-    String password,
-    int port: 3306,
-    String database,
-    String output: "dbschema.dart",
-    String library: "dbschema"}) async {
-  var config = new GeneratorConfig.fromArgs(dbType, host, user, password, port, database, output, library);
+Future<void> generate({
+  String host = "localhost",
+  String user = "root",
+  required String password,
+  int port = 3306,
+  required String database,
+  String output = "dbschema.dart",
+  String library = "dbschema",
+}) async {
+  final config = GeneratorConfig.fromArgs(
+    host: host,
+    port: port,
+    database: database,
+    output: output,
+    library: library,
+    username: user,
+    password: password,
+  );
 
-  Adapter db = new Adapter.build(config.adapterConfig);
+  final db = config.conn;
+
+  await db.open();
+
   var tableNames = (await db.query("SHOW TABLES")).map((r) => r[0]);
   var code = new StringBuffer();
   code.write("""library ${config.library};
@@ -117,7 +137,8 @@ import 'package:squilder/squilder.dart';
   for (var tableName in tableNames) {
     var results = await db.query("DESCRIBE $tableName");
     var fields = results.map((result) {
-      return new _Field(result[0], result[1].toString(), result[2] == "YES", result[3], result[4], result[5].toString());
+      return new _Field(result[0], result[1].toString(), result[2] == "YES",
+          result[3], result[4], result[5].toString());
     });
     code.writeln(_generateCodeForTable(tableName, fields));
   }
