@@ -1,6 +1,6 @@
 part of squilder.condition;
 
-class EqualFieldCondition<T> extends Object with Condition {
+class EqualFieldCondition<T> with Condition {
   final TableField<T> fieldOne;
   final TableField<T> fieldTwo;
 
