@@ -1,14 +1,14 @@
 part of squilder.condition;
 
-class GenericCondition extends Object with Condition {
+class GenericCondition with Condition {
   static final regexp = new RegExp(r"\?");
   final String string;
   final Iterable params;
 
-  GenericCondition(this.string, [this.params]);
+  GenericCondition(this.string, [this.params = const []]);
 
   String toSql() {
-    var paramsList = (params ?? []).toList();
+    var paramsList = params.toList();
     var newString = string;
     while (paramsList.isNotEmpty) {
       var param = paramsList.removeAt(0);
@@ -18,6 +18,6 @@ class GenericCondition extends Object with Condition {
   }
 }
 
-GenericCondition cond(String string, [Iterable params]) {
+GenericCondition cond(String string, [Iterable params = const []]) {
   return new GenericCondition(string, params);
 }
