@@ -1,12 +1,10 @@
 part of squilder.condition;
 
-class GreaterFieldCondition<T> extends Object with Condition {
+class GreaterFieldCondition<T> with Condition {
   final TableField<T> fieldOne;
   final TableField<T> fieldTwo;
 
   GreaterFieldCondition(this.fieldOne, this.fieldTwo);
 
-  String toSql() {
-    return "${fieldOne.toSql()} > ${fieldTwo.toSql()}";
-  }
+  String toSql() => "${fieldOne.toSql()} > ${fieldTwo.toSql()}";
 }
