@@ -6,7 +6,5 @@ class LessEqualObjectCondition<T> extends Object with Condition {
 
   LessEqualObjectCondition(this.field, this.object);
 
-  String toSql() {
-    return "${field.toSql()} <= ${utils.objectToSql(object)}";
-  }
+  String toSql() => "${field.toSql()} <= ${utils.objectToSql(object)}";
 }
