@@ -1,17 +1,31 @@
 library squilder.generator.generator_config;
 
-import 'package:dapter/dapter.dart';
+import 'package:postgres/postgres.dart';
 
 class GeneratorConfig {
-  final AdapterConfig adapterConfig;
+  final PostgreSQLConnection conn;
   final String output;
   final String library;
 
-  GeneratorConfig(this.adapterConfig, this.output, this.library);
+  GeneratorConfig._(this.conn, this.output, this.library);
 
-  factory GeneratorConfig.fromArgs(String dbType, String host, String user, String password, int port, String database, String output, String library) {
-    var adapterConfig = new AdapterConfig(dbType: dbType, host: host, user: user, password: password, port: port, database: database);
-
-    return new GeneratorConfig(adapterConfig, output, library);
-  }
+  factory GeneratorConfig.fromArgs({
+    required String host,
+    required int port,
+    required String database,
+    required String output,
+    required String library,
+    String? username,
+    String? password,
+  }) =>
+      GeneratorConfig._(
+          PostgreSQLConnection(
+            host,
+            port,
+            database,
+            password: password,
+            username: username,
+          ),
+          output,
+          library);
 }
