@@ -1,8 +1,8 @@
 library squilder.table_fields;
 
+import 'package:squilder/src/condition.dart';
 import 'package:squilder/src/serializable.dart';
 import 'package:squilder/src/table.dart';
-import 'package:squilder/src/condition.dart';
 import 'package:squilder/utils.dart' as utils;
 
 abstract class TableFields {
@@ -13,61 +13,43 @@ class TableField<T> implements Serializable {
   final String name;
   final Table table;
   final String defaultValue;
-  TableField(this.table, this.name, [this.defaultValue]);
 
-  EqualFieldCondition<T> eqToField(TableField<T> field) {
-    return new EqualFieldCondition<T>(this, field);
-  }
+  TableField(this.table, this.name, [this.defaultValue = ""]);
 
-  EqualObjectCondition<T> eqToObj(T obj) {
-    return new EqualObjectCondition<T>(this, obj);
-  }
+  EqualFieldCondition<T> eqToField(TableField<T> field) =>
+      EqualFieldCondition<T>(this, field);
 
-  LessFieldCondition<T> ltField(TableField<T> field) {
-    return new LessFieldCondition<T>(this, field);
-  }
+  EqualObjectCondition<T> eqToObj(T obj) => EqualObjectCondition<T>(this, obj);
 
-  LessObjectCondition<T> ltObj(T obj) {
-    return new LessObjectCondition<T>(this, obj);
-  }
+  LessFieldCondition<T> ltField(TableField<T> field) =>
+      LessFieldCondition<T>(this, field);
 
-  LessEqualFieldCondition<T> lteField(TableField<T> field) {
-    return new LessEqualFieldCondition<T>(this, field);
-  }
+  LessObjectCondition<T> ltObj(T obj) => LessObjectCondition<T>(this, obj);
 
-  LessEqualObjectCondition<T> lteObj(T obj) {
-    return new LessEqualObjectCondition<T>(this, obj);
-  }
+  LessEqualFieldCondition<T> lteField(TableField<T> field) =>
+      LessEqualFieldCondition<T>(this, field);
 
-  GreaterFieldCondition<T> gtField(TableField<T> field) {
-    return new GreaterFieldCondition<T>(this, field);
-  }
+  LessEqualObjectCondition<T> lteObj(T obj) =>
+      LessEqualObjectCondition<T>(this, obj);
 
-  GreaterObjectCondition<T> gtObj(T obj) {
-    return new GreaterObjectCondition<T>(this, obj);
-  }
+  GreaterFieldCondition<T> gtField(TableField<T> field) =>
+      GreaterFieldCondition<T>(this, field);
 
-  GreaterEqualFieldCondition<T> gteField(TableField<T> field) {
-    return new GreaterEqualFieldCondition<T>(this, field);
-  }
+  GreaterObjectCondition<T> gtObj(T obj) =>
+      GreaterObjectCondition<T>(this, obj);
 
-  GreaterEqualObjectCondition<T> gteObj(T obj) {
-    return new GreaterEqualObjectCondition<T>(this, obj);
-  }
+  GreaterEqualFieldCondition<T> gteField(TableField<T> field) =>
+      GreaterEqualFieldCondition<T>(this, field);
 
-  Condition like(String pattern) {
-    return new LikeCondition(this, pattern);
-  }
+  GreaterEqualObjectCondition<T> gteObj(T obj) =>
+      GreaterEqualObjectCondition<T>(this, obj);
 
-  String toSql() {
-    return "${table.toSql()}.`${utils.escape(name)}`";
-  }
+  Condition like(String pattern) => LikeCondition(this, pattern);
 
-  bool operator ==(other) {
-    return other is TableField && other.name == name && other.table == table;
-  }
+  String toSql() => "${table.toSql()}.`${utils.escape(name)}`";
+
+  bool operator ==(other) =>
+      other is TableField && other.name == name && other.table == table;
 
   int get hashCode => name.hashCode ^ table.hashCode;
 }
-
-
